Refactor ActorCriticPolicy
Now we have 2 framework classes:
TrainingPolicy: a class for policy that do training using OnPolicyAlgorithm
OnPolicyAlgorithm: a generic interface for on-policy algorithms

And implementation for Actor Critic:
ActorCriticAlgorithm: the on-policy algorithm for actor critic
ActorCriticLoss: loss class for actor critic
